Large reorganization
- Improve data API consistency
- Reorganize module naming and location
- Reorganize class naming and location
