+++ examples/strategies/ema_cross_filtered.py
@@ -123,7 +123,7 @@ class EMACrossFiltered(TradingStrategy):
         """
         Actions to be performed on strategy start.
         """
-        instrument = self.get_instrument(self.symbol)
+        instrument = self.instrument(self.symbol)
 
         self.precision = instrument.price_precision
         self.entry_buffer = instrument.tick_size.as_double() * 3.0
@@ -144,8 +144,8 @@ class EMACrossFiltered(TradingStrategy):
         self._update_news_event()
 
         # Get historical data
-        self.get_quote_ticks(self.symbol)
-        self.get_bars(self.bar_type)
+        self.request_quote_ticks(self.symbol)
+        self.request_bars(self.bar_type)
 
         # Subscribe to live data
         self.subscribe_instrument(self.symbol)
@@ -300,7 +300,7 @@ class EMACrossFiltered(TradingStrategy):
         self.unsubscribe_quote_ticks(self.symbol)
 
     def _check_signal(self, bar: Bar, sl_buffer: float, spread_buffer: float):
-        if self.count_orders_working() == 0 and self.is_flat():  # No active or pending positions
+        if self.orders_working_count() == 0 and self.is_flat():  # No active or pending positions
             # BUY LOGIC
             if self.fast_ema.value >= self.slow_ema.value:
                 self._enter_long(bar, sl_buffer, spread_buffer)
