+++ nautilus_trader/backtest/engine.pxd
@@ -23,8 +23,8 @@ from nautilus_trader.backtest.execution_client cimport BacktestExecClient
 from nautilus_trader.backtest.simulated_broker cimport SimulatedBroker
 from nautilus_trader.backtest.models cimport FillModel
 from nautilus_trader.common.clock cimport Clock
-from nautilus_trader.common.execution_database cimport ExecutionDatabase
-from nautilus_trader.common.execution_engine cimport ExecutionEngine
+from nautilus_trader.execution.database cimport ExecutionDatabase
+from nautilus_trader.execution.engine cimport ExecutionEngine
 from nautilus_trader.common.logging cimport Logger
 from nautilus_trader.common.logging cimport LoggerAdapter
 from nautilus_trader.common.portfolio cimport Portfolio
