+++ nautilus_trader/data/engine.pxd
@@ -29,7 +29,7 @@ from nautilus_trader.model.instrument cimport Instrument
 from nautilus_trader.model.tick cimport QuoteTick
 from nautilus_trader.model.tick cimport TradeTick
 from nautilus_trader.trading.strategy cimport TradingStrategy
-from nautilus_trader.common.data_client cimport DataClient
+from nautilus_trader.data.client cimport DataClient
 
 
 cdef class DataEngine:
@@ -61,6 +61,8 @@ cdef class DataEngine:
     cpdef void update_instruments(self, Venue venue) except *
     cpdef void update_instruments_all(self) except *
     cpdef void _internal_update_instruments(self, list instruments) except *
+    cpdef void request_instrument(self, Symbol symbol, callback) except *
+    cpdef void request_instruments(self, Venue venue, callback) except *
     cpdef void request_quote_ticks(
         self,
         Symbol symbol,
@@ -85,55 +87,55 @@ cdef class DataEngine:
         int limit,
         callback,
     ) except *
-    cpdef void request_instrument(self, Symbol symbol, callback) except *
-    cpdef void request_instruments(self, Venue venue, callback) except *
+    cpdef void subscribe_instrument(self, Symbol symbol, handler) except *
     cpdef void subscribe_quote_ticks(self, Symbol symbol, handler) except *
     cpdef void subscribe_trade_ticks(self, Symbol symbol, handler) except *
     cpdef void subscribe_bars(self, BarType bar_type, handler) except *
-    cpdef void subscribe_instrument(self, Symbol symbol, handler) except *
+    cpdef void unsubscribe_instrument(self, Symbol symbol, handler) except *
     cpdef void unsubscribe_quote_ticks(self, Symbol symbol, handler) except *
     cpdef void unsubscribe_trade_ticks(self, Symbol symbol, handler) except *
     cpdef void unsubscribe_bars(self, BarType bar_type, handler) except *
-    cpdef void unsubscribe_instrument(self, Symbol symbol, handler) except *
 
-# -- REGISTRATION METHODS ------------------------------------------------------------------------ #
+# -- REGISTRATION METHODS --------------------------------------------------------------------------
 
     cpdef void register_data_client(self, DataClient client) except *
     cpdef void register_strategy(self, TradingStrategy strategy) except *
+    cpdef list registered_venues(self)
 
-# -- HANDLER METHODS ----------------------------------------------------------------------------- #
+# -- HANDLER METHODS -------------------------------------------------------------------------------
 
+    cpdef void handle_instrument(self, Instrument instrument) except *
+    cpdef void handle_instruments(self, list instruments) except *
     cpdef void handle_quote_tick(self, QuoteTick tick, bint send_to_handlers=*) except *
     cpdef void handle_quote_ticks(self, list ticks) except *
     cpdef void handle_trade_tick(self, TradeTick tick, bint send_to_handlers=*) except *
     cpdef void handle_trade_ticks(self, list ticks) except *
     cpdef void handle_bar(self, BarType bar_type, Bar bar, bint send_to_handlers=*) except *
     cpdef void handle_bars(self, BarType bar_type, list bars) except *
-    cpdef void handle_instrument(self, Instrument instrument) except *
-    cpdef void handle_instruments(self, list instruments) except *
 
-# -- QUERY METHODS ------------------------------------------------------------------------------- #
+# -- QUERY METHODS ---------------------------------------------------------------------------------
 
-    cpdef list registered_venues(self)
     cpdef list subscribed_quote_ticks(self)
     cpdef list subscribed_trade_ticks(self)
     cpdef list subscribed_bars(self)
     cpdef list subscribed_instruments(self)
-    cpdef list instrument_symbols(self)
-    cpdef dict get_instruments(self)
-    cpdef Instrument get_instrument(self, Symbol symbol)
-    cpdef bint has_quote_ticks(self, Symbol symbol)
-    cpdef bint has_trade_ticks(self, Symbol symbol)
-    cpdef bint has_bars(self, BarType bar_type)
-    cpdef int quote_tick_count(self, Symbol symbol)
-    cpdef int trade_tick_count(self, Symbol symbol)
-    cpdef int bar_count(self, BarType bar_type)
+
+    cpdef list symbols(self)
+    cpdef list instruments(self)
     cpdef list quote_ticks(self, Symbol symbol)
     cpdef list trade_ticks(self, Symbol symbol)
     cpdef list bars(self, BarType bar_type)
+    cpdef Instrument instrument(self, Symbol symbol)
     cpdef QuoteTick quote_tick(self, Symbol symbol, int index=*)
     cpdef TradeTick trade_tick(self, Symbol symbol, int index=*)
     cpdef Bar bar(self, BarType bar_type, int index=*)
+    cpdef int quote_tick_count(self, Symbol symbol)
+    cpdef int trade_tick_count(self, Symbol symbol)
+    cpdef int bar_count(self, BarType bar_type)
+    cpdef bint has_quote_ticks(self, Symbol symbol)
+    cpdef bint has_trade_ticks(self, Symbol symbol)
+    cpdef bint has_bars(self, BarType bar_type)
+
     cpdef double get_exchange_rate(
         self,
         Currency from_currency,
