+++ nautilus_trader/live/execution_database.pxd
@@ -13,7 +13,7 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
-from nautilus_trader.common.execution_database cimport ExecutionDatabase
+from nautilus_trader.execution.database cimport ExecutionDatabase
 from nautilus_trader.serialization.base cimport CommandSerializer
 from nautilus_trader.serialization.base cimport EventSerializer
