+++ nautilus_trader/network/queue.pyx
@@ -15,8 +15,8 @@
 
 import threading
 
-from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.common.logging cimport LoggerAdapter
+from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.network.socket cimport Socket
