+++ nautilus_trader/trading/trader.pxd
@@ -16,8 +16,8 @@
 from nautilus_trader.analysis.performance cimport PerformanceAnalyzer
 from nautilus_trader.analysis.reports cimport ReportProvider
 from nautilus_trader.common.clock cimport Clock
-from nautilus_trader.common.data_engine cimport DataEngine
-from nautilus_trader.common.execution_engine cimport ExecutionEngine
+from nautilus_trader.data.engine cimport DataEngine
+from nautilus_trader.execution.engine cimport ExecutionEngine
 from nautilus_trader.common.logging cimport LoggerAdapter
 from nautilus_trader.common.portfolio cimport Portfolio
 from nautilus_trader.common.uuid cimport UUIDFactory
