+++ tests/unit_tests/data/test_aggregation.py
@@ -17,16 +17,13 @@ from datetime import datetime
 from datetime import timedelta
 import unittest
 
-from pandas import Timestamp
 import pytz
 
 from nautilus_trader.backtest.clock import TestClock
 from nautilus_trader.backtest.logging import TestLogger
-from nautilus_trader.common.market import BarBuilder
-from nautilus_trader.common.market import BarDataWrangler
-from nautilus_trader.common.market import TickBarAggregator
-from nautilus_trader.common.market import TickDataWrangler
-from nautilus_trader.common.market import TimeBarAggregator
+from nautilus_trader.data.aggregation import BarBuilder
+from nautilus_trader.data.aggregation import TickBarAggregator
+from nautilus_trader.data.aggregation import TimeBarAggregator
 from nautilus_trader.model.bar import BarSpecification
 from nautilus_trader.model.bar import BarType
 from nautilus_trader.model.enums import BarAggregation
@@ -34,124 +31,11 @@ from nautilus_trader.model.enums import PriceType
 from nautilus_trader.model.objects import Price
 from nautilus_trader.model.objects import Quantity
 from nautilus_trader.model.tick import QuoteTick
-from tests.test_kit.data import TestDataProvider
 from tests.test_kit.mocks import ObjectStorer
 from tests.test_kit.stubs import TestStubs
 from tests.test_kit.stubs import UNIX_EPOCH
 
 AUDUSD_FXCM = TestStubs.symbol_audusd_fxcm()
-
-
-class TickDataWranglerTests(unittest.TestCase):
-
-    def setUp(self):
-        self.clock = TestClock()
-
-    def test_tick_data(self):
-        # Arrange
-        # Act
-        ticks = TestDataProvider.usdjpy_test_ticks()
-
-        # Assert
-        self.assertEqual(1000, len(ticks))
-
-    def test_build_with_tick_data(self):
-        # Arrange
-        tick_data = TestDataProvider.usdjpy_test_ticks()
-        bid_data = TestDataProvider.usdjpy_1min_bid()
-        ask_data = TestDataProvider.usdjpy_1min_ask()
-        self.tick_builder = TickDataWrangler(
-            instrument=TestStubs.instrument_usdjpy(),
-            data_ticks=tick_data,
-            data_bars_bid={BarAggregation.MINUTE: bid_data},
-            data_bars_ask={BarAggregation.MINUTE: ask_data})
-
-        # Act
-        self.tick_builder.build(0)
-        ticks = self.tick_builder.tick_data
-
-        # Assert
-        self.assertEqual(BarAggregation.TICK, self.tick_builder.resolution)
-        self.assertEqual(1000, len(ticks))
-        self.assertEqual(Timestamp("2013-01-01 22:02:35.907000", tz="UTC"), ticks.iloc[1].name)
-
-    def test_build_ticks_with_bar_data(self):
-        # Arrange
-        bid_data = TestDataProvider.usdjpy_1min_bid()
-        ask_data = TestDataProvider.usdjpy_1min_ask()
-        self.tick_builder = TickDataWrangler(
-            instrument=TestStubs.instrument_usdjpy(),
-            data_ticks=None,
-            data_bars_bid={BarAggregation.MINUTE: bid_data},
-            data_bars_ask={BarAggregation.MINUTE: ask_data})
-
-        # Act
-        self.tick_builder.build(0)
-        tick_data = self.tick_builder.tick_data
-
-        # Assert
-        self.assertEqual(BarAggregation.MINUTE, self.tick_builder.resolution)
-        self.assertEqual(1491252, len(tick_data))
-        self.assertEqual(Timestamp("2013-01-01T21:59:59.900000+00:00", tz="UTC"), tick_data.iloc[0].name)
-        self.assertEqual(Timestamp("2013-01-01T21:59:59.900000+00:00", tz="UTC"), tick_data.iloc[1].name)
-        self.assertEqual(Timestamp("2013-01-01T21:59:59.900000+00:00", tz="UTC"), tick_data.iloc[2].name)
-        self.assertEqual(Timestamp("2013-01-01T22:00:00.000000+00:00", tz="UTC"), tick_data.iloc[3].name)
-        self.assertEqual(0, tick_data.iloc[0]["symbol"])
-        self.assertEqual(0, tick_data.iloc[0]["bid_size"])
-        self.assertEqual(0, tick_data.iloc[0]["ask_size"])
-        self.assertEqual(0, tick_data.iloc[1]["bid_size"])
-        self.assertEqual(0, tick_data.iloc[1]["ask_size"])
-        self.assertEqual(0, tick_data.iloc[2]["bid_size"])
-        self.assertEqual(0, tick_data.iloc[2]["ask_size"])
-        self.assertEqual(1.5, tick_data.iloc[3]["bid_size"])
-        self.assertEqual(2.25, tick_data.iloc[3]["ask_size"])
-
-
-class BarDataWranglerTests(unittest.TestCase):
-
-    def setUp(self):
-        data = TestDataProvider.gbpusd_1min_bid()[:1000]
-        self.bar_builder = BarDataWrangler(5, 1, data)
-
-    def test_can_build_bars_all(self):
-        # Arrange
-        # Act
-        bars = self.bar_builder.build_bars_all()
-
-        # Assert
-        self.assertEqual(1000, len(bars))
-
-    def test_can_build_bars_range_with_defaults(self):
-        # Arrange
-        # Act
-        bars = self.bar_builder.build_bars_range()
-
-        # Assert
-        self.assertEqual(999, len(bars))
-
-    def test_can_build_bars_range_with_param(self):
-        # Arrange
-        # Act
-        bars = self.bar_builder.build_bars_range(start=500)
-
-        # Assert
-        self.assertEqual(499, len(bars))
-
-    def test_can_build_bars_from_with_defaults(self):
-        # Arrange
-        # Act
-        bars = self.bar_builder.build_bars_from()
-
-        # Assert
-        self.assertEqual(1000, len(bars))
-
-    def test_can_build_bars_from_with_param(self):
-        # Arrange
-        # Act
-        bars = self.bar_builder.build_bars_from(index=500)
-
-        # Assert
-        self.assertEqual(500, len(bars))
 
 
 class BarBuilderTests(unittest.TestCase):
